+++ Libplanet.Tests/Blockchain/Policies/VolatileStagePolicyTest.cs
@@ -1,0 +1,14 @@
+using Libplanet.Blockchain.Policies;
+using Libplanet.Tests.Common.Action;
+
+namespace Libplanet.Tests.Blockchain.Policies
+{
+    public class VolatileStagePolicyTest : StagePolicyTest
+    {
+        private readonly VolatileStagePolicy<DumbAction> _stagePolicy =
+            new VolatileStagePolicy<DumbAction>();
+
+        protected override IStagePolicy<DumbAction> StagePolicy =>
+            _stagePolicy;
+    }
+}