+++ Libplanet/Blockchain/BlockChain.cs
@@ -28,9 +28,8 @@ namespace Libplanet.Blockchain
     /// A class have <see cref="Block{T}"/>s, <see cref="Transaction{T}"/>s, and the chain
     /// information.
     /// <para>In order to watch its state changes, implement <see cref="IRenderer{T}"/>
-    /// interface and pass it to the <see
-    /// cref="BlockChain{T}(IBlockPolicy{T},IStore,IStateStore,Block{T},IEnumerable{IRenderer{T}})"
-    /// /> constructor.</para>
+    /// interface and pass it to the <see cref="BlockChain{T}(IBlockPolicy{T}, IStagePolicy{T},
+    /// IStore, IStateStore, Block{T}, IEnumerable{IRenderer{T}})"/> constructor.</para>
     /// </summary>
     /// <remarks>This object is guaranteed that it has at least one block, since it takes a genesis
     /// block when it's instantiated.</remarks>
@@ -74,6 +73,7 @@ namespace Libplanet.Blockchain
         /// </summary>
         /// <param name="policy"><see cref="IBlockPolicy{T}"/> to use in the
         /// <see cref="BlockChain{T}"/>.</param>
+        /// <param name="stagePolicy">The staging policy to follow.</param>
         /// <param name="store"><see cref="IStore"/> to store <see cref="Block{T}"/>s,
         /// <see cref="Transaction{T}"/>s, and <see cref="BlockChain{T}"/> information.</param>
         /// <param name="genesisBlock">The genesis <see cref="Block{T}"/> of
@@ -90,13 +90,15 @@ namespace Libplanet.Blockchain
         /// (i.e., <paramref name="genesisBlock"/>).</exception>
         public BlockChain(
             IBlockPolicy<T> policy,
+            IStagePolicy<T> stagePolicy,
             IStore store,
             IStateStore stateStore,
             Block<T> genesisBlock,
             IEnumerable<IRenderer<T>> renderers = null
-            )
+        )
             : this(
                 policy,
+                stagePolicy,
                 store,
                 stateStore,
                 store.GetCanonicalChainId() ?? Guid.NewGuid(),
@@ -108,6 +110,7 @@ namespace Libplanet.Blockchain
 
         internal BlockChain(
             IBlockPolicy<T> policy,
+            IStagePolicy<T> stagePolicy,
             IStore store,
             IStateStore stateStore,
             Guid id,
@@ -116,6 +119,7 @@ namespace Libplanet.Blockchain
         )
             : this(
                 policy,
+                stagePolicy,
                 store,
                 stateStore,
                 id,
@@ -128,6 +132,7 @@ namespace Libplanet.Blockchain
 
         private BlockChain(
             IBlockPolicy<T> policy,
+            IStagePolicy<T> stagePolicy,
             IStore store,
             IStateStore stateStore,
             Guid id,
@@ -138,6 +143,7 @@ namespace Libplanet.Blockchain
         {
             Id = id;
             Policy = policy;
+            StagePolicy = stagePolicy;
             Store = store;
 
             // It expects store is DefaultStore or RocksDBStore.
@@ -223,7 +229,8 @@ namespace Libplanet.Blockchain
         /// <remarks>
         /// Since this value is immutable, renderers cannot be registered after once a <see
         /// cref="BlockChain{T}"/> object is instantiated; use <c>renderers</c> option of <see cref=
-        /// "BlockChain{T}(IBlockPolicy{T},IStore,IStateStore,Block{T},IEnumerable{IRenderer{T}})"/>
+        /// "BlockChain{T}(IBlockPolicy{T}, IStagePolicy{T}, IStore, IStateStore, Block{T},
+        /// IEnumerable{IRenderer{T}})"/>
         /// constructor instead.
         /// </remarks>
         public IImmutableList<IRenderer<T>> Renderers { get; }
@@ -234,7 +241,15 @@ namespace Libplanet.Blockchain
         /// </summary>
         public IImmutableList<IActionRenderer<T>> ActionRenderers { get; }
 
+        /// <summary>
+        /// The block and blockchain policy.
+        /// </summary>
         public IBlockPolicy<T> Policy { get; }
+
+        /// <summary>
+        /// The staging policy.
+        /// </summary>
+        public IStagePolicy<T> StagePolicy { get; set; }
 
         /// <summary>
         /// The topmost <see cref="Block{T}"/> of the current blockchain.
@@ -441,6 +456,11 @@ namespace Libplanet.Blockchain
         /// <see cref="Transaction{T}"/> with a given <paramref name="txId"/>.</exception>
         public Transaction<T> GetTransaction(TxId txId)
         {
+            if (StagePolicy.Get(this, txId, includeUnstaged: true) is { } tx)
+            {
+                return tx;
+            }
+
             _rwlock.EnterReadLock();
             try
             {
@@ -615,21 +635,9 @@ namespace Libplanet.Blockchain
                     msg);
             }
 
-            // FIXME it's global chain lock so using it in this method can cause degrading
-            // parallelism of `BlockChain<T>`. we should re-organize locks in `BlockChain<T>`
-            _rwlock.EnterWriteLock();
-
-            try
+            if (!StagePolicy.HasStaged(this, transaction.Id, includeUnstaged: true))
             {
-                if (!_transactions.ContainsKey(transaction.Id))
-                {
-                    _transactions[transaction.Id] = transaction;
-                    Store.StageTransactionIds(ImmutableHashSet.Create(transaction.Id));
-                }
-            }
-            finally
-            {
-                _rwlock.ExitWriteLock();
+                StagePolicy.Stage(this, transaction);
             }
         }
 
@@ -639,21 +647,8 @@ namespace Libplanet.Blockchain
         /// <param name="transaction">A <see cref="Transaction{T}"/>
         /// to remove from the pending list.</param>
         /// <seealso cref="StageTransaction"/>
-        public void UnstageTransaction(Transaction<T> transaction)
-        {
-            // FIXME it's global chain lock so using it in this method can cause degrading
-            // parallelism of `BlockChain<T>`. we should re-organize locks in `BlockChain<T>`
-            _rwlock.EnterWriteLock();
-
-            try
-            {
-                Store.UnstageTransactionIds(ImmutableHashSet.Create(transaction.Id));
-            }
-            finally
-            {
-                _rwlock.ExitWriteLock();
-            }
-        }
+        public void UnstageTransaction(Transaction<T> transaction) =>
+            StagePolicy.Unstage(this, transaction.Id);
 
         /// <summary>
         /// Gets next <see cref="Transaction{T}.Nonce"/> of the address.
@@ -664,34 +659,24 @@ namespace Libplanet.Blockchain
         /// <paramref name="address"/>.</returns>
         public long GetNextTxNonce(Address address)
         {
-            _rwlock.EnterReadLock();
+            long nonce = Store.GetTxNonce(Id, address);
+            long prevNonce = nonce - 1;
+            IOrderedEnumerable<long> stagedTxNonces = StagePolicy.Iterate(this)
+                .Where(tx => tx.Signer.Equals(address) && tx.Nonce > prevNonce)
+                .Select(tx => tx.Nonce)
+                .OrderBy(n => n);
 
-            try
+            foreach (long n in stagedTxNonces)
             {
-                long nonce = Store.GetTxNonce(Id, address);
-                var prevNonce = nonce - 1;
-                var stagedTxNonces = Store.IterateStagedTransactionIds()
-                    .Select(Store.GetTransaction<T>)
-                    .Where(tx => tx.Signer.Equals(address) && tx.Nonce > prevNonce)
-                    .Select(tx => tx.Nonce)
-                    .OrderBy(n => n);
-
-                foreach (long n in stagedTxNonces)
+                if (n != nonce)
                 {
-                    if (n != nonce)
-                    {
-                        break;
-                    }
-
-                    nonce++;
+                    break;
                 }
 
-                return nonce;
+                nonce++;
             }
-            finally
-            {
-                _rwlock.ExitReadLock();
-            }
+
+            return nonce;
         }
 
 #pragma warning disable MEN003
@@ -1061,16 +1046,8 @@ namespace Libplanet.Blockchain
         /// <returns><see cref="IImmutableSet{TxId}"/> of staged transactions.</returns>
         public IImmutableSet<TxId> GetStagedTransactionIds()
         {
-            _rwlock.EnterReadLock();
-
-            try
-            {
-                return Store.IterateStagedTransactionIds().ToImmutableHashSet();
-            }
-            finally
-            {
-                _rwlock.ExitReadLock();
-            }
+            // FIXME: How about turning this method to the StagedTransactions property?
+            return StagePolicy.Iterate(this).Select(tx => tx.Id).ToImmutableHashSet();
         }
 
         internal void Append(
@@ -1163,6 +1140,7 @@ namespace Libplanet.Blockchain
                     nonceDeltas[txSigner] = nonceDelta + 1;
                 }
 
+                ImmutableDictionary<Address, long> maxNonces;
                 _rwlock.EnterWriteLock();
                 try
                 {
@@ -1181,7 +1159,7 @@ namespace Libplanet.Blockchain
 
                     _logger.Debug("Unstaging transactions...");
 
-                    ImmutableDictionary<Address, long> maxNonces = block.Transactions
+                    maxNonces = block.Transactions
                         .GroupBy(
                             t => t.Signer,
                             t => t.Nonce,
@@ -1192,20 +1170,24 @@ namespace Libplanet.Blockchain
                             }
                         )
                         .ToImmutableDictionary(t => t.signer, t => t.maxNonce);
-                    ISet<TxId> txIds = Store.IterateStagedTransactionIds()
-                        .Select(Store.GetTransaction<T>)
-                        .Where(tx => maxNonces.TryGetValue(tx.Signer, out long nonce) &&
-                            tx.Nonce <= nonce)
-                        .Select(tx => tx.Id)
-                        .ToImmutableHashSet();
-                    Store.UnstageTransactionIds(txIds);
-                    TipChanged?.Invoke(this, (prevTip, block));
-                    _logger.Debug("Block {blockIndex}: {block} is appended.", block?.Index, block);
                 }
                 finally
                 {
                     _rwlock.ExitWriteLock();
                 }
+
+                ISet<TxId> txIds = StagePolicy.Iterate(this)
+                    .Where(tx => maxNonces.TryGetValue(tx.Signer, out long nonce) &&
+                        tx.Nonce <= nonce)
+                    .Select(tx => tx.Id)
+                    .ToImmutableHashSet();
+                foreach (TxId txId in txIds)
+                {
+                    StagePolicy.Unstage(this, txId);
+                }
+
+                TipChanged?.Invoke(this, (prevTip, block));
+                _logger.Debug("Block {blockIndex}: {block} is appended.", block?.Index, block);
 
                 if (renderBlocks)
                 {
@@ -1489,7 +1471,7 @@ namespace Libplanet.Blockchain
                 ? Renderers
                 : Enumerable.Empty<IRenderer<T>>();
             var forked = new BlockChain<T>(
-                Policy, Store, StateStore, Guid.NewGuid(), Genesis, true, renderers);
+                Policy, StagePolicy, Store, StateStore, Guid.NewGuid(), Genesis, true, renderers);
             Guid forkedId = forked.Id;
             _logger.Debug(
                 "Trying to fork chain at {branchPoint}" +
@@ -1728,19 +1710,22 @@ namespace Libplanet.Blockchain
                 _rwlock.EnterWriteLock();
                 try
                 {
-                    IEnumerable<TxId>
-                        GetTxIdsWithRange(BlockChain<T> chain, Block<T> start, Block<T> end)
+                    IEnumerable<Transaction<T>>
+                        GetTxsWithRange(BlockChain<T> chain, Block<T> start, Block<T> end)
                         => Enumerable
                             .Range((int)start.Index + 1, (int)(end.Index - start.Index))
-                            .SelectMany(x => chain[x].Transactions.Select(tx => tx.Id));
+                            .SelectMany(x => chain[x].Transactions);
 
                     // It assumes reorg is small size. If it was big, this may be heavy task.
-                    ImmutableHashSet<TxId> unstagedTxIds =
-                        GetTxIdsWithRange(this, topmostCommon, Tip).ToImmutableHashSet();
-                    ImmutableHashSet<TxId> stageTxIds =
-                        GetTxIdsWithRange(other, topmostCommon, other.Tip).ToImmutableHashSet();
-                    ImmutableHashSet<TxId> restageTxIds = unstagedTxIds.Except(stageTxIds);
-                    Store.StageTransactionIds(restageTxIds);
+                    ImmutableHashSet<Transaction<T>> unstagedTxs =
+                        GetTxsWithRange(this, topmostCommon, Tip).ToImmutableHashSet();
+                    ImmutableHashSet<Transaction<T>> stageTxs =
+                        GetTxsWithRange(other, topmostCommon, other.Tip).ToImmutableHashSet();
+                    ImmutableHashSet<Transaction<T>> restageTxs = unstagedTxs.Except(stageTxs);
+                    foreach (Transaction<T> restageTx in restageTxs)
+                    {
+                        StagePolicy.Stage(this, restageTx);
+                    }
 
                     Guid obsoleteId = Id;
                     Id = other.Id;
@@ -1801,10 +1786,7 @@ namespace Libplanet.Blockchain
 
         internal ImmutableArray<Transaction<T>> ListStagedTransactions()
         {
-            Transaction<T>[] txs = Store
-                .IterateStagedTransactionIds()
-                .Select(Store.GetTransaction<T>)
-                .ToArray();
+            Transaction<T>[] txs = StagePolicy.Iterate(this).ToArray();
 
             Dictionary<Address, LinkedList<Transaction<T>>> seats = txs
                 .GroupBy(tx => tx.Signer)